refactor(web): move AI API types to shared layer for FSD compliance
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
